+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Nicer error message on the error page
 - Add mail threading to mission control
 - Setting to ask participants if it's their first time
+- Setting for maximum allowed points earned total
 
 ### Changed
 - Participant mail subject to be clickable